% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{call_synthdid}
\alias{call_synthdid}
\title{Call Synthdid Estimation Function}
\usage{
call_synthdid(dta, outcome_colname, estimator, covariate_colnames)
}
\arguments{
\item{dta}{One of the champion data sets bundled in the package (\code{\link{lol_champ_pool_dta}} or \code{\link{lol_champ_dta}}).}

\item{outcome_colname}{Name of the column of \code{lol_champ_pool_dta} storing the outcome of interest.}

\item{estimator}{Which estimator to use. Must be one of "sc" (standard synthetic control), "sc_reg" (sc + a ridge penalty), "synthdid" (synthetic diff-in-diff).}

\item{covariate_colnames}{Character vector with the names of the columns storing the time-varying covariates for which we want to adjust for. If empty, no adjustment is performed. If non-empty, we adjust the outcome using the \code{xsynthdid} package.}
}
\description{
Calls one of the \code{synthdid} estimation functions.
}
\details{
Before calling the estimation function, \code{\link{call_synthdid}} processes \code{dta} to generate the required objects.\cr

If \code{covariate_colnames} is non-empty, \code{\link{call_synthdid}} regresses the outcomes on the covariates, on time fixed effects, and unit fixed effects, with the regression estimated without
using observations in which the treatment takes place. Then, \code{\link{call_synthdid}} uses the estimated coefficients to compute the residuals of the outcomes for all observations (also those in which
the treatment takes place), and uses these residuals rather than the original outcomes for estimation.
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
