% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{champions_performance_plots_lol}
\alias{champions_performance_plots_lol}
\title{LoL Performance Plots}
\usage{
champions_performance_plots_lol(
  champions,
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-09-12"),
  bandwidth = 0.01,
  save_here = getwd()
)
}
\arguments{
\item{champions}{Character vector with the champions of interest.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}

\item{bandwidth}{Parameter controlling the amount of smoothing.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces a nice plot.
}
\description{
Produces plots for kill-to-death ratio, number of assists, number of kills, and win rates of the champions of interest.
}
\details{
\code{treatment_date1}, \code{treatment_date2}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

The series are smoothed using a Nadaraya–Watson kernel regression. The user can control the amount of smoothing by setting the \code{bandwidth} parameter. The larger parameter, the smoother the series.
An infinitesimal bandwidth amounts to no smoothing.\cr
}
\seealso{
\code{\link{descriptive_plots_lol}} \code{\link{descriptive_plots_lol_lgb}}
}
\author{
Riccardo Di Francesco
}
