% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-construct-data.R
\name{clean_lol_data}
\alias{clean_lol_data}
\title{Cleaning LoL Data}
\usage{
clean_lol_data(dta)
}
\arguments{
\item{dta}{Data set as constructed by the \code{\link{pull_lol_data}} function (you can rowbind more data sets and pass in the final one).}
}
\description{
Cleans the raw LoL data as obtained from calling the \code{\link{pull_lol_data}} function.
}
\details{
\code{\link{clean_lol_data}} performs the following operations on the raw data.\cr

First, it keeps data for the period January-July 2022.

Second, it aggregates the regions of interest according to the geographical macro area where servers hosting the match are located (Latin America, North America, Europe, and Korea)
and drops regions with too many missing values (Oceania, Russia, Japan, and Turkey).\cr

Third, it drops duplicated matches and matches with corrupted data (i.e., matches with less than ten players or which lasted more than two hours or where the position of some player is not assigned).\cr

Fourth, it hard-codes the 'ban' variable (it maps it from integer numbers to champions' names). To apply the function to more recent years, new champions must be hard-coded.\cr

Fifth, it identifies and assign the main and the auxiliary positions of each champion based on where they are played the most. It then drops matches where champions were played in weird roles.\cr

The raw data are available from the authors on request.
}
\seealso{
\code{\link{pull_lol_data}}, \code{\link{construct_lol_champion_data}}, \code{\link{construct_lol_champion_pooled_data}}, \code{\link{construct_lol_player_data}}
}
\author{
Riccardo Di Francesco
}
