% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{construct_donor_pool}
\alias{construct_donor_pool}
\title{Construct Donor Pool}
\usage{
construct_donor_pool(dta, donor_pool, my_champion)
}
\arguments{
\item{dta}{One of the champion data sets bundled in the package (\code{\link{lol_champ_pool_dta}} or \code{\link{lol_champ_dta}}).}

\item{donor_pool}{Which units to include in the donor pool. Must be either "all" or "non_lgb". The latter excludes Nami, Leona, Diana, and Neeko.}

\item{my_champion}{Champion under investigation. We need to include it as well.}
}
\description{
Subsets the provided data to include only the champion under investigation and the desired units in the donor pool.
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
