% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid-helpers.R
\name{construct_synth_outcome}
\alias{construct_synth_outcome}
\title{Synthetic Outcome}
\usage{
construct_synth_outcome(
  object,
  dta,
  unit_colname,
  outcome_colname,
  time_colname
)
}
\arguments{
\item{object, }{An \code{\link[synthdid]{synthdid_estimate}} object.}

\item{dta}{Data set used to estimate \code{object}.}

\item{unit_colname}{Name of the column of \code{dta} storing the units of observation.}

\item{outcome_colname}{Name of the column of \code{dta} storing the outcome of interest.}

\item{time_colname}{Name of the column of \code{dta} storing the time variable.}
}
\value{
A list with the following elements:
\describe{
\item{weights}{An (nw)x2 tibble (with nw = number of non-zero estimated weights), with one column for the identities and one column for the contributions of units in the donor pool with a non-zero estimated weight.}
\item{synth_outcome}{An Tx2 tibble, with one column for the time periods and one column for the synthetic outcome.}
}
}
\description{
Constructs the synthetic outcome using the estimation results stored in an \code{\link[synthdid]{synthdid_estimate}} object.
}
\author{
Riccardo Di Francesco
}
