% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{debiased_did}
\alias{debiased_did}
\title{Debiased Machine Learning for Difference-in-Differences}
\usage{
debiased_did(Y, D, t, X, K = 2, learner = "lasso", repeat_n = 100)
}
\arguments{
\item{Y}{Outcome variable.}

\item{D}{Indicator for treated units.}

\item{t}{Indicator for post-treatment.}

\item{X}{Matrix of control variables.}

\item{K}{Controls K in K-fold cross-fitting.}

\item{learner}{String, controls the machine learner used to estimate the nuisance functions.}

\item{repeat_n}{Number of replications.}
}
\value{
Point estimate for the ATT and its standard error.
}
\description{
Fits the DMLDiD estimator.
}
\details{
To obtain a robust result, the estimator is fitted \code{repeat_n} times and the average is returned.\cr

Original code: \href{https://github.com/NengChiehChang/Diff-in-Diff/blob/master/RCS_Algorithm.R}{https://github.com/NengChiehChang/Diff-in-Diff/blob/master/RCS_Algorithm.R}.
}
\keyword{internal}
