% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{descriptive_plots_lol}
\alias{descriptive_plots_lol}
\title{LoL Descriptive Plots}
\usage{
descriptive_plots_lol(
  champions,
  lol_champ_dta,
  lol_champ_pool_dta,
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2022-07-11")
)
}
\arguments{
\item{champions}{Character vector with the champions of interest.}

\item{lol_champ_dta}{Data set bundled in the package.}

\item{lol_champ_pool_dta}{Data set bundled in the package.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}
}
\description{
Produces plots for the pick, ban, and win variables of the champions of interest.
}
\seealso{
\code{\link{descriptive_plots_lol_lgb}} \code{\link{performance_plots_lol}}
}
\author{
Riccardo Di Francesco
}
