% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{descriptive_plots_lol_lgb}
\alias{descriptive_plots_lol_lgb}
\title{LoL Outcome Plots (LGB Aggregate)}
\usage{
descriptive_plots_lol_lgb(
  lol_champ_pool_dta,
  max_date = as.POSIXct("2022-07-15")
)
}
\arguments{
\item{lol_champ_pool_dta}{Data set bundled in the package.}

\item{max_date}{Object of class \code{POSIXct}. Where to cut the series.}
}
\description{
Produces plots for the pick, ban, and win variables of the LGB aggregate, constructed by averaging variables for the LGB champions
'Nami', 'Leona', 'Diana', and 'Neeko'.
}
\seealso{
\code{\link{descriptive_plots_lol}}
}
\author{
Riccardo Di Francesco
}
