% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{descriptive_plots_lol_lgb_composite}
\alias{descriptive_plots_lol_lgb_composite}
\title{LoL Descriptive Plots (LGB Aggregate)}
\usage{
descriptive_plots_lol_lgb_composite(
  treatment_date1 = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d"),
  treatment_date2 = as.POSIXct("2023-06-01", tryFormats = "\%Y-\%m-\%d"),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-09-12"),
  save_here = getwd()
)
}
\arguments{
\item{treatment_date1}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{treatment_date2}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces nice plots.
}
\description{
Produces plots for the pick, ban, and win variables of the LGB composite unit, constructed by averaging variables for the LGB champions
'Nami', 'Leona', 'Diana', and 'Neeko'.
}
\details{
\code{treatment_date1}, \code{treatment_date2}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.
}
\seealso{
\code{\link{descriptive_plots_lol}} \code{\link{performance_plots_lol}}
}
\author{
Riccardo Di Francesco
}
