% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism.R
\name{did_players_performance}
\alias{did_players_performance}
\title{LoL Impact on Players Performance}
\usage{
did_players_performance(
  n_pre_matches,
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d")
)
}
\arguments{
\item{treatment_date}{Object of class \code{POSIXct}. The date of the treatment.}

\item{filter}{Which players to retain for the analysis.}
}
\value{
Returns a list with \code{treatment_date} and all the diff-in-diff results. The user can post-process the output using the \code{\link{plot_did}} function.
}
\description{
Uses a diff-in-diff strategy to investigate the impact of the coming-out event on players' performance.
}
\details{
We consider only "prior" users.
We define two versions of the treatment, as described in the paper. We keep the control group fixed as it is always composed of those prior-users that do not reduce their pick rates for Graves at all.\cr

The estimators of Callaway and Sant’Anna (2021) are employed to estimate the impact of the coming-out event on the performance of treated players. This is implemented using the \code{\link[did]{att_gt}} function.
Technical details are given in the associated documentation. To summarize, we identify and estimate the average treatment effect on the treated for all time t > \code{treatment_date}. "Effects" before that date are
also estimated and are useful to check the plausibility of the parallel trend assumption.\cr

\code{treatment_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr
}
\author{
Riccardo Di Francesco
}
