% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lol_champ_pool_dta}
\alias{lol_champ_pool_dta}
\title{LoL Champion Data Set (Pooled)}
\format{
For documentation, see \code{\link{construct_lol_champion_pooled_data}}.
}
\usage{
lol_champ_pool_dta
}
\description{
LoL champion data set, obtained by calling the \code{\link{construct_lol_champion_pooled_data}} function on the output of the
\code{\link{construct_lol_champion_data}} function.
}
\keyword{datasets}
