% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lol_player_dta}
\alias{lol_player_dta}
\title{LoL Player Data Set}
\format{
For documentation, see \code{\link{construct_lol_player_data}}.
}
\usage{
lol_player_dta
}
\description{
LoL champion data set, obtained by calling the \code{\link{construct_lol_player_data}} function on the output of the
\code{\link{clean_lol_data}} function.
}
\keyword{datasets}
