% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-plots.R
\name{mechanisms_plots_lol}
\alias{mechanisms_plots_lol}
\title{LoL Descriptive Plots}
\usage{
mechanisms_plots_lol(
  n_groups,
  n_pre_matches,
  n_post_matches,
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d"),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-08-01"),
  save_here = getwd()
)
}
\arguments{
\item{n_groups}{How many groups to form.}

\item{n_pre_matches}{How many matches before \code{treatment_date} players must have played to be kept in the data set.}

\item{n_post_matches}{How many matches before \code{treatment_date} players must have played to be kept in the data set.}

\item{treatment_date}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces nice plots.
}
\description{
Produces plots
}
\details{
\code{treatment_date}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

Groups are formed by cutting the Graves' pre-treatment pick rates distribution into \code{n_groups} buckets.\cr

Players that have played less than \code{n_pre_matches} before \code{treatment_date} or less than \code{n_post_matches} between \code{treatment_date} and \code{max_date} are dropped. The number of
players remaining in the data set is printed.\cr
}
\author{
Riccardo Di Francesco
}
