% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-plots.R
\name{outcome_plots_lol_lgb}
\alias{outcome_plots_lol_lgb}
\title{LoL Outcome Plots (LGB Aggregate)}
\usage{
outcome_plots_lol_lgb(lol_champ_pool_dta, max_date = as.POSIXct("2022-07-15"))
}
\arguments{
\item{lol_champ_pool_dta}{Data set as constructed by the \code{\link{construct_lol_champion_pooled_data}} function (you can find this data set already bundled in the package).}

\item{max_date}{Object of class \code{POSIXct}. Where to cut the series.}
}
\description{
Produces plots for the pick, ban, and win variables of the LGB aggregate, constructed by averaging variables for the LGB champions
'Nami', 'Leona', 'Diana', and 'Neeko'.
}
\seealso{
\code{\link{pull_lol_data}}, \code{\link{construct_lol_champion_data}}, \code{\link{construct_lol_champion_pooled_data}}, \code{\link{\code{\link{construct_lol_player_data}}}}
}
\author{
Riccardo Di Francesco
}
