% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{players_descriptive_plots_lol}
\alias{players_descriptive_plots_lol}
\title{LoL Players' Behavior Plots}
\usage{
players_descriptive_plots_lol(save_here = getwd())
}
\arguments{
\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces a nice plot.
}
\description{
Produces histograms for the number of matches and hours played.
}
\details{
This routine computes, for each player, the total number of matches and hours played, as well as daily averages of number of matches and hours played. It then displays the distributions
of these variables. Note that all variables are at the player-level.

\code{treatment_date} and must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr
}
\author{
Riccardo Di Francesco
}
