% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism.R
\name{players_performance_plots_lol}
\alias{players_performance_plots_lol}
\title{LoL Players' Performance Plots}
\usage{
players_performance_plots_lol(
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d"),
  save_here = getwd()
)
}
\arguments{
\item{treatment_date}{Object of class \code{POSIXct}. When the treatment took place.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces nice plots.
}
\description{
Divides players into Graves prior and non-prior users and produces plots showing the average performance of players
in each group before and after the treatment. It also produces a plot to investigate which positions are the most played in each group before and after the treatment.
}
\details{
\code{treatment_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr
}
\author{
Riccardo Di Francesco
}
