% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism.R
\name{plot_did}
\alias{plot_did}
\title{Diff-in-Diff Plots}
\usage{
plot_did(
  att_gt_results,
  title,
  subtitle,
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d")
)
}
\arguments{
\item{att_gt_results}{One of the results of \code{\link{did_players_performance}}.}

\item{title}{String, title of the plot.}

\item{subtitle}{String, subtitle of the plot.}

\item{treatment_date}{Object of class \code{POSIXct}. The date of the treatment. Must be the same used when calling \code{\link{did_players_performance}} for the result to make sense.}
}
\value{
Returns a nice plot.
}
\description{
Plots the time ATTs estimated by \code{\link{did_players_performance}}.
}
\author{
Riccardo Di Francesco
}
