% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{produce_latex}
\alias{produce_latex}
\title{Cost of Coming Out - Main Analysis}
\usage{
produce_latex(pooled_result_list, regional_result_list)
}
\arguments{
\item{pooled_result_list}{A list storing the output of \code{\link{run_main_pooled}}. Each element must store the results obtained with a different donor pool or estimator. Everything else must be equal for the table to make sense.}

\item{regional_result_list}{A list storing the output of \code{\link{run_main_regional}}. Each element must store the results obtained with a different donor pool or estimator. Everything else must be equal for the table to make sense.}
}
\value{
Prints LATEX code for a nice table.
}
\description{
Produced LATEX code for a table displaying the results of the main analysis of the Cost of Coming Out paper performed by \code{\link{run_main_pooled}} and \code{\link{run_main_regional}}.
}
\details{
Different synthetic control specifications (e.g., donor pool, estimator) can be used when calling \code{\link{run_main_pooled}} and \code{\link{run_main_regional}}. \code{pooled_result_list} and
\code{regional_result_list} must be lists with each element storing the results from a single specification. Please make sure that the lists use the specification order. Failure to comply will raise an error.\cr

The order of the columns of the table will follow the order of the lists.\cr

While some parameters can change, others must be fixed over the specifications. These include the treatment date, the champion assigned to the treatment, the outcome series, and the bandwidths.
}
\seealso{
\code{\link{produce_plots_pooled}} \code{\link{produce_plots_regional}} \code{\link{produce_latex_pooled}}
}
\author{
Riccardo Di Francesco
}
