% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{produce_latex_regional}
\alias{produce_latex_regional}
\title{Cost of Coming Out - Main Analysis}
\usage{
produce_latex_regional(regional_result_list)
}
\arguments{
\item{regional_result_list}{A list storing the output of \code{\link{run_main_regional}}. Each element must store the results obtained with a different donor pool. Everything else must be equal for the table to make sense.}
}
\value{
Prints LATEX code for a nice table.
}
\description{
Produced LATEX code for a table displaying the results of the main analysis of the Cost of Coming Out paper performed by \code{\link{run_main_regional_result_list}}.
}
\details{
Different synthetic control specifications can be used by calling \code{\link{run_main_pooled}} multiple times in a for loop, with some parameters (e.g., donor pool, estimator)
changing over the iterations. \code{regional_result_list} must be a list storing the results from a single iteration. Please order the loop according to the order you want to display in the table.\cr

While some parameters can change, others must be fixed over the specifications. These include the treatment date, the champion assigned to the treatment, the outcome series, and the bandwidths.
}
\seealso{
\code{\link{produce_plots_pooled}} \code{\link{produce_plots_regional}} \code{\link{produce_latex_pooled}}
}
\author{
Riccardo Di Francesco
}
