% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid-helpers.R
\name{produce_plot_placebo}
\alias{produce_plot_placebo}
\title{Placebo Plot}
\usage{
produce_plot_placebo(
  pooled_results,
  main_champion,
  to_plot_n,
  ylims = c(0, 100),
  save_here = getwd()
)
}
\arguments{
\item{pooled_results}{Output of \code{\link{run_main_pooled}} called on several distinct characters.}

\item{main_champion}{String denoting the actual treated champion.}

\item{to_plot_n}{Plot only results for the characters with the lowest pre-treatment RMSE.}

\item{ylims}{Vector storing lower and upper limit for the y-axis.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Save some nice plots and returns a tibble storing the pre-treatment root mean squared errors of all fits.
}
\description{
Produced a placebo plot for synthetic control estimators.
}
\seealso{
\code{\link{produce_plots_regional}} \code{\link{produce_latex_pooled}} \code{\link{produce_latex_regional}}
}
\author{
Riccardo Di Francesco
}
