% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{produce_plots_regional}
\alias{produce_plots_regional}
\title{Cost of Coming Out - Main Analysis}
\usage{
produce_plots_regional(regional_results)
}
\arguments{
\item{regional_results}{Output of \code{\link{run_main_regional}}.}
}
\value{
Save some nice plots.
}
\description{
Runs the main analysis of the Cost of Coming Out paper using \code{\link{lol_champ_dta}}.
}
\seealso{
\code{\link{produce_plots_pooled}}
}
\author{
Riccardo Di Francesco
}
