% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-data.R
\name{pull_lol_data}
\alias{pull_lol_data}
\title{Pull LoL Data}
\usage{
pull_lol_data(server, queue, rank, division, api_key, min_sec, max_sec)
}
\arguments{
\item{server}{ID of the server of interest.}

\item{queue}{Queue type.}

\item{rank}{Rank of interest.}

\item{division}{Division of interest.}

\item{api_key}{Your API key. It must follow the format \code{"RGAPI-xxxx"}.}

\item{min_sec}{Wait for a random amount of seconds bounded from below by \code{min_sec} before pulling next item.}

\item{max_sec}{Wait for a random amount of seconds bounded from above by \code{max_sec} before pulling next item.}
}
\value{
Prints a csv file storing a data set with the following columns:
\describe{
  \item{\code{match_id}}{Match ID.}
  \item{\code{day}}{Time and date of the beginning of the match.}
  \item{\code{duration}}{Duration of the match in minutes.}
  \item{\code{queue_id}}{Queue ID. See \href{https://static.developer.riotgames.com/docs/lol/queues.json}{here}. Currently, only 5v5 Ranked Solo games are returned.}
  \item{\code{game_mode}}{Game mode. See \href{https://static.developer.riotgames.com/docs/lol/gameModes.json}{here}. Currently, only \code{"CLASSIC"} matches are returned.}
  \item{\code{game_type}}{Game type. See \href{https://static.developer.riotgames.com/docs/lol/gameTypes.json}{here}. Currently, only \code{"MATCHED_GAME"} matches are returned.}
  \item{\code{player_name}}{Name of each player in the match.}
  \item{\code{player_id}}{ID of each player in the match.}
  \item{\code{player_puuid}}{PUIID of each player in the match.}
  \item{\code{player_level}}{Level of each player in the match.}
  \item{\code{champion}}{Champion picked by each player in the match.}
  \item{\code{position}}{Position played by each player in the match.}
  \item{\code{kills}}{Number of kills of each player in the match.}
  \item{\code{assists}}{Number of assists of each player in the match.}
  \item{\code{deaths}}{Number of deaths of each player in the match.}
  \item{\code{gold}}{Gold earned by each player in the match.}
  \item{\code{early_surrender}}{Whether the match ended by early surrender.}
  \item{\code{surrender}}{Whether the match ended by surrender.}
  \item{\code{win}}{Whether the player belongs to the winning team.}
  \item{\code{ban}}{Champions banned for a given game.}
}
}
\description{
Pulls raw data from Riot API.
}
\details{
The following are the only valid values for the inputs:
\describe{
  \item{\code{server}}{\code{"br1"}, \code{"eun1"}, \code{"euw1"}, \code{"jp1"}, \code{"kr"}, \code{"la1"}, \code{"la2"}, \code{"na1"}, \code{"oc1"}, \code{"ru"}, \code{"tr1"}}
  \item{\code{queue}}{\code{"RANKED_SOLO_5x5"}, \code{"RANKED_TFT"}, \code{"RANKED_FLEX_SR"}, \code{"RANKED_FLEX_TT"}}
  \item{\code{rank}}{\code{"CHALLENGER"}, \code{"GRANDMASTER"}, \code{"MASTER"}, \code{"DIAMOND"}, \code{"PLATINUM"}, \code{"GOLD"}, \code{"SILVER"}, \code{"BRONZE"}, \code{"IRON"}}
  \item{\code{division}}{\code{"I"}, \code{"II"}, \code{"III"}, \code{"IV"}}
}

To get an API key, you need a Riot account of at least level 5. Then, visit \href{https://developer.riotgames.com/}{https://developer.riotgames.com/} to generate a key.\cr

For more details on the output, please refer to Riot's official documentation:
\href{https://developer.riotgames.com/apis#summoner-v4/GET_getBySummonerName}{https://developer.riotgames.com/apis#summoner-v4/GET_getBySummonerName},
\href{https://developer.riotgames.com/apis#match-v5/GET_getMatchIdsByPUUID}{https://developer.riotgames.com/apis#match-v5/GET_getMatchIdsByPUUID}, and
\href{https://developer.riotgames.com/apis#match-v5/GET_getMatch}{https://developer.riotgames.com/apis#match-v5/GET_getMatch}.
}
\seealso{
\code{\link{pull_lol_match_ids}}, \code{\link{pull_lol_match_info}}
}
\author{
Riccardo Di Francesco
}
