% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-data.R
\name{pull_lol_match_ids}
\alias{pull_lol_match_ids}
\title{Pulling LoL Match IDs}
\usage{
pull_lol_match_ids(server, player_name, api_key)
}
\arguments{
\item{server}{ID of the server of interest.}

\item{player_name}{The player's nickname.}

\item{api_key}{Your API key. It must follow the format \code{"RGAPI-xxxx"}.}
}
\value{
A vector of matches id recently played by \code{summoner_name}.\cr
}
\description{
Pulls the match IDs of a given player.
}
\details{
The following are the valid values for the inputs:
\describe{
  \item{\code{server}}{\code{"br1"}, \code{"eun1"}, \code{"euw1"}, \code{"jp1"}, \code{"kr"}, \code{"la1"}, \code{"la2"}, \code{"na1"}, \code{"oc1"}, \code{"ru"}, \code{"tr1"}}
}

To get an API key, you need a Riot account of at least level 5. Then, visit \href{https://developer.riotgames.com/}{https://developer.riotgames.com/} to generate a key.\cr
}
\seealso{
\code{\link{pull_lol_players}}, \code{\link{pull_lol_match_info}}, \code{\link{pull_lol_data}}
}
\author{
Riccardo Di Francesco
}
