% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-data.R
\name{pull_lol_match_info}
\alias{pull_lol_match_info}
\title{Pulling LoL Match Information}
\usage{
pull_lol_match_info(server, match_id, api_key, min_sec = 0, max_sec = 0)
}
\arguments{
\item{server}{ID of the server of interest.}

\item{match_id}{The match ID of interest.}

\item{api_key}{Your API key. It must follow the format \code{"RGAPI-xxxx"}.}

\item{min_sec}{Wait for a random amount of seconds bounded from below by \code{min_sec} before pulling next item.}

\item{max_sec}{Wait for a random amount of seconds bounded from above by \code{max_sec} before pulling next item.}
}
\value{
A data frame with the following variables:
\describe{
  \item{\code{match_id}}{Match ID.}
  \item{\code{day}}{Time and date of the beginning of the match.}
  \item{\code{duration}}{Duration of the match in minutes.}
  \item{\code{queue_id}}{Queue ID. See \href{https://static.developer.riotgames.com/docs/lol/queues.json}{here}. Currently, only 5v5 Ranked Solo games are returned.}
  \item{\code{game_mode}}{Game mode. See \href{https://static.developer.riotgames.com/docs/lol/gameModes.json}{here}. Currently, only \code{"CLASSIC"} matches are returned.}
  \item{\code{game_type}}{Game type. See \href{https://static.developer.riotgames.com/docs/lol/gameTypes.json}{here}. Currently, only \code{"MATCHED_GAME"} matches are returned.}
  \item{\code{player_name}}{Name of each player in the match.}
  \item{\code{player_id}}{ID of each player in the match.}
  \item{\code{player_puuid}}{PUIID of each player in the match.}
  \item{\code{player_level}}{Level of each player in the match.}
  \item{\code{champion}}{Champion picked by each player in the match.}
  \item{\code{position}}{Position played by each player in the match.}
  \item{\code{kills}}{Number of kills of each player in the match.}
  \item{\code{assists}}{Number of assists of each player in the match.}
  \item{\code{deaths}}{Number of deaths of each player in the match.}
  \item{\code{gold}}{Gold earned by each player in the match.}
  \item{\code{early_surrender}}{Whether the match ended by early surrender.}
  \item{\code{surrender}}{Whether the match ended by surrender.}
  \item{\code{win}}{Whether the player belongs to the winning team.}
  \item{\code{ban}}{Champions banned for a given game.}
}

For more details on these variables, please refer to Riot's official documentation:
\href{https://developer.riotgames.com/apis#summoner-v4/GET_getBySummonerName}{https://developer.riotgames.com/apis#summoner-v4/GET_getBySummonerName},
\href{https://developer.riotgames.com/apis#match-v5/GET_getMatchIdsByPUUID}{https://developer.riotgames.com/apis#match-v5/GET_getMatchIdsByPUUID}, and
\href{https://developer.riotgames.com/apis#match-v5/GET_getMatch}{https://developer.riotgames.com/apis#match-v5/GET_getMatch}.
}
\description{
Pulls information about the match of interest.
}
\details{
The following are the valid values for the inputs:
\describe{
  \item{\code{server}}{\code{"br1"}, \code{"eun1"}, \code{"euw1"}, \code{"jp1"}, \code{"kr"}, \code{"la1"}, \code{"la2"}, \code{"na1"}, \code{"oc1"}, \code{"ru"}, \code{"tr1"}}
}

Only classic games are returned.\cr

To get an API key, you need a Riot account of at least level 5. Then, visit \href{https://developer.riotgames.com/}{https://developer.riotgames.com/} to generate a key.\cr

Imposing a waiting time using \code{min_sec} and \code{max_sec} between each request is useful to avoid hitting the rate limits imposed by Riot. With a personal API key, we can execute 20 requests
every 1 second and 100 requests every 2 minutes. To use a deterministic amount of time, set \code{min_sec} and \code{max_sec} to the same number.
}
\seealso{
\code{\link{pull_lol_players}}, \code{\link{pull_lol_match_ids}}, \code{\link{pull_lol_data}}
}
\author{
Riccardo Di Francesco
}
