% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-data.R
\name{pull_lol_players}
\alias{pull_lol_players}
\title{Pulling LoL Players}
\usage{
pull_lol_players(server, queue, rank, division, api_key)
}
\arguments{
\item{server}{ID of the server of interest.}

\item{queue}{Queue type.}

\item{rank}{Rank of interest.}

\item{division}{Division of interest.}

\item{api_key}{Your API key. It must follow the format \code{"RGAPI-xxxx"}.}
}
\description{
Pulls names and IDs of all the players currently belonging to a given rank and division.
}
\details{
The following are the only valid values for the inputs:
\describe{
  \item{\code{server}}{\code{"br1"}, \code{"eun1"}, \code{"euw1"}, \code{"jp1"}, \code{"kr"}, \code{"la1"}, \code{"la2"}, \code{"na1"}, \code{"oc1"}, \code{"ru"}, \code{"tr1"}}
  \item{\code{queue}}{\code{"RANKED_SOLO_5x5"}, \code{"RANKED_TFT"}, \code{"RANKED_FLEX_SR"}, \code{"RANKED_FLEX_TT"}}
  \item{\code{rank}}{\code{"CHALLENGER"}, \code{"GRANDMASTER"}, \code{"MASTER"}, \code{"DIAMOND"}, \code{"PLATINUM"}, \code{"GOLD"}, \code{"SILVER"}, \code{"BRONZE"}, \code{"IRON"}}
  \item{\code{division}}{\code{"I"}, \code{"II"}, \code{"III"}, \code{"IV"}}
}

To get an API key, you need a Riot account of at least level 5. Then, visit \href{https://developer.riotgames.com/}{https://developer.riotgames.com/} to generate a key.\cr
}
\seealso{
\code{\link{pull_lol_match_ids}}, \code{\link{pull_lol_match_info}}, \code{\link{pull_lol_data}}
}
\author{
Riccardo Di Francesco
}
