% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis.R
\name{run_main_pooled}
\alias{run_main_pooled}
\title{Cost of Coming Out - Main Analysis}
\usage{
run_main_pooled(
  dta,
  champions,
  outcome_colname,
  donor_pool,
  estimator,
  treatment_date,
  backdate,
  covariate_colnames = c(),
  max_date = as.POSIXct("2022-07-15")
)
}
\arguments{
\item{dta}{Data set \code{\link{lol_champ_pool_dta}} bundled in the package.}

\item{champions}{Character vector with the names of the champions of interest. The routine performs the analysis for all them separately.}

\item{outcome_colname}{Name of the column of \code{lol_champ_pool_dta} storing the outcome of interest.}

\item{donor_pool}{Which units to include in the donor pool. Must be either "all" or "non_lgb". The latter excludes Graves, Nami, Leona, Diana, and Neeko.}

\item{estimator}{Which estimator to use. Must be one of "sc" (standard synthetic control), "sc_reg" (sc + a ridge penalty), "synthdid" (synthetic diff-in-diff).}

\item{treatment_date}{When the treatment took place. Must be of class \code{as.POSIXct} with \code{tryFormats} set to "\%Y-\%m-\%d".}

\item{backdate}{How many periods to backdate the treatment for a robustness check.}

\item{covariate_colnames}{Character vector with the names of the columns of \code{lol_champ_pool_dta} storing the time-varying covariates for which we want to adjust for. If empty, no adjustment is performed. If non-empty, we adjust the outcome using the \code{xsynthdid} package.}

\item{max_date}{Object of class \code{POSIXct}. Where to cut the series.}
}
\value{
A list of the same length as \code{champions}. Each element stores a list with results for one of those champions (elements are named).
Each inner list stores three elements, one for the results of main fit, one for the results of the backdating exercise, and one for the results of the leave-one-out exercise.
The outer list stores two additional last elements, one storing \code{outcome_colname} and one storing \code{dta} up to \code{max_date}.
}
\description{
Runs the main analysis of the Cost of Coming Out paper using \code{\link{lol_champ_pool_dta}}.
}
\details{
For each champion in \code{champions}, \code{\link{run_main_pooled}} performs the following operations.

First, it constructs the treatment variable, which equals one for the champion under investigation starting from \code{treatment_date}.\cr

Second, it subsets \code{\link{lol_champ_pool_dta}} to include only the champion under investigation and the desired units in the donor pool (as controlled by \code{donor_pool}).\cr

Third, it calls one of the \code{synthdid} estimation functions, as controlled by \code{estimator}. If \code{covariate_colnames} is non-empty, \code{\link{run_main_pooled}} regresses the outcomes
on the covariates, on time fixed effects, and unit fixed effects, with the regression estimated without using observations in which the treatment takes place. Then, \code{\link{run_main_pooled}} uses
the estimated coefficients to compute the residuals of the outcomes for all observations (also those in which the treatment takes place), and use these residuals rather than the original outcomes for
estimation.\cr

Fourth, it calls again one of the \code{synthdid} estimation functions, this time using the backdated treatment.\cr

Fifth, it performs a leave-one-out exercise, where we repeatedly estimate the synthetic control series by excluding one champion with non-zero estimated weights at a time from the donor pool. This exercise
is carried out only if the number of non-zero contributions is less than 10 to avoid a large running time.\cr

It is possible to include \code{"LGB"} in \code{champions}. If so, \code{\link{run_main_pooled}} constructs a new unit by averaging \code{outcome_colname} of Nami, Leona, Diana, and Neeko and
runs the analysis detailed above on this new unit. This is compatible only with \code{donor_pool} set to \code{"non_lgb"}.
}
\seealso{
\code{\link{run_main_regional}}
}
\author{
Riccardo Di Francesco
}
