% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis.R
\name{run_main_regional}
\alias{run_main_regional}
\title{Cost of Coming Out - Main Analysis by Region}
\usage{
run_main_regional(
  champions,
  outcome_colname,
  donors,
  estimator,
  treatment_date,
  backdate,
  inference = FALSE,
  n_boot = 100,
  bandwidth = 0.01,
  covariate_colnames = c(),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-09-12")
)
}
\arguments{
\item{champions}{Character vector with the names of the champions of interest. The routine performs the analysis for all them separately.}

\item{outcome_colname}{Name of the column of \code{\link{lol_champ_dta}} storing the outcome of interest.}

\item{donors}{Which units to include in the donor pool. See the details section below.}

\item{estimator}{Which estimator to use. Must be one of "sc" (standard synthetic control), "sc_reg" (sc plus a ridge penalty), "synthdid" (synthetic diff-in-diff).}

\item{treatment_date}{Object of class \code{POSIXct}. When the treatment took place.}

\item{backdate}{How many periods to backdate the treatment for a robustness check.}

\item{inference}{Logical, whether to estimate standard errors. If \code{TRUE}, the placebo method described in Section 5 of Arkhangelsky et al. is used.}

\item{n_boot}{Number of champions to be assigned the placebo treatment for standard error estimation. Ignored if \code{inference} is \code{FALSE}.}

\item{bandwidth}{Parameter controlling the amount of smoothing.}

\item{covariate_colnames}{Character vector with the names of the columns of \code{\link{lol_champ_dta}} storing the time-varying covariates for which we want to adjust for.}

\item{min_date}{Object of class \code{POSIXct}. When to start the series.}

\item{max_date}{Object of class \code{POSIXct}. When to end the series.}
}
\value{
A list of the same length as \code{champions}. Each element stores a list with results for one of those champions (elements are named).
Each inner list stores two elements: a list with the results of the main fit for each region, and a list of regional data sets used for estimation.
The outer list stores five additional elements: the name of the outcome, the name of the estimator, the name of the donors, the treatment date, and the bandwidth used to smooth the outcome series.
}
\description{
Runs the main analysis of the Cost of Coming Out paper using \code{\link{lol_champ_dta}}.
}
\details{
\code{\link{run_main_regional}} disaggregates \code{\link{lol_champ_dta}} by constructing four data sets, one for each region. Then, for each champion in \code{champions}, it performs
the following operations separately for each data set.\cr

\itemize{
 \item{First: }{constructs the treatment variable, which equals one for the champion under investigation starting from \code{treatment_date}.}
 \item{Second: }{subsets each regional data set to include only the champion under investigation and the desired units in the donor pool (as controlled by \code{donors}).}
 \item{Third: }{calls one of the \code{synthdid} estimation functions, as controlled by \code{estimator}. If \code{covariate_colnames} is non-empty, \code{\link{run_main_pooled}} regresses the outcomes
                on the covariates, on time fixed effects, and unit fixed effects, with the regression estimated without using observations in which the treatment takes place. Then, \code{\link{run_main_pooled}} uses
                the estimated coefficients to compute the residuals of the outcomes for all observations (also those in which the treatment takes place), and use these residuals rather than the original outcomes for
                estimation.}
}

\code{treatment_date}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

The outcome series is smoothed using a Nadaraya–Watson kernel regression before the covariate adjustment. The user can control the amount of smoothing by setting the \code{bandwidth} parameter. The larger parameter, the smoother the series.
An infinitesimal bandwidth amounts to no smoothing.\cr

\code{donors} must be a character vector with one or more champions contained in \code{dta}. Alternatively, one of seven special strings can be used:

\itemize{
   \item{"all" }{This includes all champions.}
   \item{"non_lgb" }{This includes all champions expect Graves, Twisted Fate, Diana, Leona, Nami, and Neeko.}
   \item{"jungle" }{This includes all champions whose main role is Jungle.}
   \item{"middle" }{This includes all champions whose main role is Middle.}
   \item{"top" }{This includes all champions whose main role is Top.}
   \item{"support" }{This includes all champions whose main role is Support.}
   \item{"adc" }{This includes all champions whose main role is Adc.}
   \item{"main_role"}{This includes all champions whose main role is the same as the champion of interest.}
   \item{"aux_role"}{This includes all champions whose auxiliary role is the same as the champion of interest.}
}

Almost all of these choices (exceptions are \code{"all"} and \code{"non_lgb"}) can be combined with the prefix \code{"non"} to exclude champions from a particular role.
For instance, setting \code{donors} to \code{non_jungle} excludes from the donor pool all champions whose main role is jungle.

It is possible to include \code{"LGB"} in \code{champions}. If so, \code{\link{run_main_pooled}} constructs a new unit by averaging the outcomes of Diana, Leona, Nami, and Neeko and
runs the analysis detailed above on this new unit. This is compatible only with \code{donor_pool} set to \code{"non_lgb"}.
}
\seealso{
\code{\link{run_main_pooled}}
}
\author{
Riccardo Di Francesco
}
