% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_split}
\alias{sample_split}
\title{Sample Splitting}
\usage{
sample_split(n, K = 2)
}
\arguments{
\item{n}{Sample size.}

\item{K}{Number of folds.}
}
\value{
A list with indexes for the first and the second fold.
}
\description{
Spits the sample into K folds of roughly the same size.
}
\details{
Currently, only \code{K = 2} is supported.
}
\keyword{internal}
