% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid-helpers.R
\name{sc_estimate_reg}
\alias{sc_estimate_reg}
\title{Regularized Synthetic Control}
\usage{
sc_estimate_reg(Y, N0, T0, X = array(dim = c(dim(Y), 0)))
}
\arguments{
\item{Y}{Observation matrix.}

\item{N0}{Number of control units (control units are in the rows 1-N0 of Y).}

\item{T0}{Number of pre-treatment periods (they correspond to columns 1-T0 of Y).}

\item{X}{Optional 3D array of time-varying covariates. Shape should be N X T X C for C covariates.}
}
\description{
Synthetic control with a ridge penalty, as defined in Algorithm 1 of Arkhangelsky et al. (2021). Taken from
\href{https://synth-inference.github.io/synthdid/articles/paper-results.html}{https://synth-inference.github.io/synthdid/articles/paper-results.html}.
}
\author{
Riccardo Di Francesco
}
