% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sleep}
\alias{sleep}
\title{Waiting for Next Iteration}
\usage{
sleep(min_sec = 0, max_sec = 0)
}
\arguments{
\item{min_sec}{Minimum amount of seconds to be waited.}

\item{max_sec}{Maximum amount of seconds to be waited.}
}
\description{
It stops the execution of R for a random number of seconds. Useful to avoid hitting rate limits while web-scraping or pulling from APIs.
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
