% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{champions_descriptive_plots_lol}
\alias{champions_descriptive_plots_lol}
\title{LoL Descriptive Plots}
\usage{
champions_descriptive_plots_lol(
  champions,
  treatment_date1 = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d"),
  treatment_date2 = as.POSIXct("2023-06-01", tryFormats = "\%Y-\%m-\%d"),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-09-12"),
  bandwidth = 0.01,
  save_here = getwd()
)
}
\arguments{
\item{champions}{Character vector with the champions of interest.}

\item{treatment_date1}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{treatment_date2}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}

\item{bandwidth}{Parameter controlling the amount of smoothing.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces nice plots.
}
\description{
Produces plots for the pick, ban, and win variables of the champions of interest.
}
\details{
\code{treatment_date1}, \code{treatment_date2}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

\code{\link{champions_descriptive_plots_lol}} always includes in the plots an LGB composite unit constructed by averaging the variables values of the champions
Nami, Leona, Diana, and Neeko.\cr

The series are smoothed using a Nadaraya–Watson kernel regression. The user can control the amount of smoothing by setting the \code{bandwidth} parameter. The larger parameter, the smoother the series.
An infinitesimal bandwidth amounts to no smoothing.\cr
}
\seealso{
\code{\link{descriptive_plots_lol_lgb}} \code{\link{performance_plots_lol}}
}
\author{
Riccardo Di Francesco
}
