% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid-helpers.R
\name{construct_donor_pool}
\alias{construct_donor_pool}
\title{Construct Donor Pool}
\usage{
construct_donor_pool(dta, donors, my_champion)
}
\arguments{
\item{dta}{One of the champion data sets bundled in the package (\code{\link{lol_champ_pool_dta}} or \code{\link{lol_champ_dta}}).}

\item{donors}{Which units to include in the donor pool. See the details section below.}

\item{my_champion}{Champion under investigation. We need to include it as well.}
}
\value{
\code{dta} filtered to contain only the champion under investigation and the desired donors.
}
\description{
Subsets the provided data to include only the champion under investigation and the desired units in the donor pool.
}
\details{
\code{donors} must be a character vector with one or more champions contained in \code{dta}. Alternatively, one of seven special strings can be used:

\itemize{
   \item{"all" }{This includes all champions.}
   \item{"non_lgb" }{This includes all champions expect Graves, Nami, Leona, Diana, and Neeko.}
   \item{"jungle" }{This includes all champions whose main role is Jungle.}
   \item{"middle" }{This includes all champions whose main role is Middle.}
   \item{"top" }{This includes all champions whose main role is Top.}
   \item{"support" }{This includes all champions expect whose main role is Support.}
   \item{"adc" }{This includes all champions expect whose main role is Adc.}
}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
