% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-construct-data.R
\name{construct_lol_player_data}
\alias{construct_lol_player_data}
\title{Construct LoL Player Data Set}
\usage{
construct_lol_player_data(dta)
}
\arguments{
\item{dta}{Data set as constructed by the \code{\link{clean_lol_data}} function.}
}
\description{
Constructs the LoL player data set.
}
\details{
\code{construct_lol_player_data} performs the following operations on \code{dta}.\cr

First, it keeps only data for the year 2022.\cr

Second, it generates the variables of interest.
\describe{
 \item{\code{Graves}}{How often the player picked Graves in that \code{day}.}
 \item{\code{Graves_ban}}{How often the player banned Graves in that \code{day}.}
 \item{\code{win_sum}}{How many matches the player won in that \code{day}.}
 \item{\code{n_matches}}{How many matches the player has played in that \code{day}.}
 \item{\code{Graves_rate}}{Equals \code{Graves} divided by \code{n_matches}.}
 \item{\code{Graves_ban_rate}}{Equals \code{Graves_ban} divided by \code{n_matches}.}
 \item{\code{win_rate}}{Equals \code{win_sum} divided by \code{n_matches}.}
 \item{\code{gold_avg}}{Average gold earned by the player in that \code{day}.}
 \item{\code{kills_avg}}{Average number of kills achieved by the player in that \code{day}.}
 \item{\code{assists_avg}}{Average number of assists achieved by the player in that \code{day}.}
 \item{\code{deaths_avg}}{Average number of deaths achieved by the player in that \code{day}.}
 \item{\code{day_no}}{Numeric version of the \code{day} variable, useful when the data set is loaded in other sessions to get the time right.}
}
}
\seealso{
\code{\link{pull_lol_data}}, \code{\link{clean_lol_data}}, \code{\link{construct_lol_champion_data}} \code{\link{construct_lol_champion_pooled_data}}
}
\author{
Riccardo Di Francesco
}
