% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{descriptive_plots_lol}
\alias{descriptive_plots_lol}
\title{LoL Descriptive Plots}
\usage{
descriptive_plots_lol(
  champions,
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d"),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-09-12")
)
}
\arguments{
\item{champions}{Character vector with the champions of interest.}

\item{treatment_date}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}
}
\description{
Produces plots for the pick, ban, and win variables of the champions of interest.
}
\details{
\code{treatment_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

\code{min_date} and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD")}.
}
\seealso{
\code{\link{descriptive_plots_lol_lgb}} \code{\link{performance_plots_lol}}
}
\author{
Riccardo Di Francesco
}
