% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism.R
\name{did_players_performance}
\alias{did_players_performance}
\title{LoL Impact on Players Performance}
\usage{
did_players_performance(
  n_groups,
  n_pre_matches,
  n_post_matches,
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d"),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-08-01")
)
}
\arguments{
\item{n_groups}{How many groups to form.}

\item{n_pre_matches}{How many matches before \code{treatment_date} players must have played to be kept in the data set.}

\item{n_post_matches}{How many matches before \code{treatment_date} players must have played to be kept in the data set.}

\item{treatment_date}{Object of class \code{POSIXct}. Where to display a dashed vertical line. Set to \code{NULL} if you do not want this line.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}
}
\value{
Prints code for a LATEX table.
}
\description{
Uses a diff-in-diff strategy to investigate the impact of the coming-out event on players' performance.
}
\details{
For each player, \code{\link{did_players_performance}} computes the average kills-to-deaths ratios and win rates before and after the coming-out event. It then
groups players by cutting the Graves' pre-treatment pick rates distribution into \code{n_groups} buckets. The largest groups (that is, those who used to
select Graves quite often before his disclosure) are considered "treated" after the occurence of the coming-out event. Two-way fixed effect regressions are employed
to estimate the impact of the coming-out event on the players' performance.\cr

\code{treatment_date}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

Players that have played less than \code{n_pre_matches} before \code{treatment_date} or less than \code{n_post_matches} between \code{treatment_date} and \code{max_date} are dropped. The number of
players remaining in the data set is printed.
}
\author{
Riccardo Di Francesco
}
