% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism.R
\name{did_players_performance}
\alias{did_players_performance}
\title{LoL Impact on Players Performance}
\usage{
did_players_performance(
  n_pre_matches,
  treatment_date = as.POSIXct("2022-06-01", tryFormats = "\%Y-\%m-\%d")
)
}
\arguments{
\item{n_pre_matches}{How many matches before \code{treatment_date} players must have played to be kept in the data set.}

\item{treatment_date}{Object of class \code{POSIXct}. The date of the treatment.}

\item{filter}{Which players to retain for the analysis.}
}
\value{
Returns a list with \code{treatment_date} and all the diff-in-diff results. The user can post-process the output using the \code{\link{plot_did}} function.
}
\description{
Uses a diff-in-diff strategy to investigate the impact of the coming-out event on players' performance.
}
\details{

}
\author{
Riccardo Di Francesco
}
