% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-plots.R
\name{players_descriptive_plots_lol}
\alias{players_descriptive_plots_lol}
\title{LoL Players' Behavior Plots}
\usage{
players_descriptive_plots_lol(save_here = getwd())
}
\arguments{
\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Produces a nice plot.
}
\description{
Produces plots for the number of matches and hours played.
}
\details{
\code{treatment_date1} and must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr
}
\author{
Riccardo Di Francesco
}
