% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism.R
\name{plot_did}
\alias{plot_did}
\title{Diff-in-Diff Plots}
\usage{
plot_did(did_results, save_here = getwd())
}
\arguments{
\item{did_results}{The results of \code{\link{did_players_performance}}.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Returns a nice plot.
}
\description{
Plots the time ATTs estimated by \code{\link{did_players_performance}}.
}
\author{
Riccardo Di Francesco
}
