% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{produce_plots_pooled}
\alias{produce_plots_pooled}
\title{Cost of Coming Out - Main Analysis}
\usage{
produce_plots_pooled(pooled_results)
}
\arguments{
\item{pooled_results}{Output of \code{\link{run_main_pooled}}.}
}
\value{
Save some nice plots.
}
\description{
Runs the main analysis of the Cost of Coming Out paper using \code{\link{lol_champ_pool_dta}}.
}
\seealso{
\code{\link{produce_plots_regional}}
}
\author{
Riccardo Di Francesco
}
