% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis-helpers.R
\name{produce_plots_regional}
\alias{produce_plots_regional}
\title{Cost of Coming Out - Main Analysis}
\usage{
produce_plots_regional(regional_results, save_here = getwd())
}
\arguments{
\item{regional_results}{Output of \code{\link{run_main_regional}}.}

\item{save_here}{String denoting the path where to save the figures.}
}
\value{
Save some nice plots.
}
\description{
Produced plots displaying the results of the main analysis of the Cost of Coming Out paper performed by \code{\link{run_main_regional}}.
}
\seealso{
\code{\link{produce_plots_pooled}}
}
\author{
Riccardo Di Francesco
}
