% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-analysis.R
\name{run_main_pooled}
\alias{run_main_pooled}
\title{Cost of Coming Out - Main Analysis}
\usage{
run_main_pooled(
  champions,
  outcome_colname,
  donors,
  estimator,
  treatment_date,
  backdate,
  inference = FALSE,
  bandwidth = 0.01,
  covariate_colnames = c(),
  min_date = as.POSIXct("2022-01-01"),
  max_date = as.POSIXct("2023-09-12")
)
}
\arguments{
\item{champions}{Character vector with the names of the champions of interest. The routine performs the analysis for all them separately.}

\item{outcome_colname}{Name of the column of \code{\link{lol_champ_pool_dta}} storing the outcome of interest.}

\item{donors}{Which units to include in the donor pool. See the details section below.}

\item{estimator}{Which estimator to use. Must be one of "sc" (standard synthetic control), "sc_reg" (sc plus a ridge penalty), "synthdid" (synthetic diff-in-diff).}

\item{treatment_date}{Object of class \code{POSIXct}. When the treatment took place.}

\item{backdate}{How many periods to backdate the treatment for a robustness check.}

\item{inference}{Logical, whether to estimate standard errors. If \code{TRUE}, the placebo method described in Section 5 of Arkhangelsky et al. is used.}

\item{bandwidth}{Parameter controlling the amount of smoothing.}

\item{covariate_colnames}{Character vector with the names of the columns of \code{\link{lol_champ_pool_dta}} storing the time-varying covariates for which we want to adjust for.}

\item{min_date}{Object of class \code{POSIXct}. Where to start the series.}

\item{max_date}{Object of class \code{POSIXct}. Where to end the series.}
}
\value{
A list of the same length as \code{champions}. Each element stores a list with results for one of those champions (elements are named).
Each inner list stores four elements: the results of the main fit, the results of the backdating exercise, the results of the leave-one-out exercise, and the data used for estimation.
The outer list stores five additional elements: the name of the outcome, the name of the estimator, the name of the donors, the treatment date, and the bandwidth used to smooth the outcome series.
}
\description{
Runs the main analysis of the Cost of Coming Out paper using \code{\link{lol_champ_pool_dta}}.
}
\details{
For each champion in \code{champions}, \code{\link{run_main_pooled}} performs the following operations.

\itemize{
 \item{First: }{constructs the treatment variable, which equals one for the champion under investigation starting from \code{treatment_date}.}
 \item{Second: }{subsets \code{\link{lol_champ_pool_dta}} to include only the champion under investigation and the desired units in the donor pool (as controlled by \code{donors}).}
 \item{Third: }{calls one of the \code{synthdid} estimation functions, as controlled by \code{estimator}. If \code{covariate_colnames} is non-empty, \code{\link{run_main_pooled}} regresses the outcomes
               on the covariates, on time fixed effects, and unit fixed effects, with the regression estimated without using observations in which the treatment takes place. Then, \code{\link{run_main_pooled}} uses
               the estimated coefficients to compute the residuals of the outcomes for all observations (also those in which the treatment takes place), and use these residuals rather than the original outcomes for
               estimation.}
 \item{Fourth: }{calls again one of the \code{synthdid} estimation functions, this time using the backdated treatment.}
 \item{Fifth: }{performs a leave-one-out exercise, where we repeatedly estimate the synthetic control series by excluding one champion with non-zero estimated weights at a time from the donor pool. This exercise
               is carried out only if the number of non-zero contributions is less than 10 to avoid a large running time.}
}

\code{treatment_date}, \code{min_date}, and \code{max_date} must be created by \code{as.POSIXct("YYYY-MM-DD", tryFormats = "\%Y-\%m-\%d")}.\cr

The outcome series is smoothed using a Nadaraya–Watson kernel regression before the covariate adjustment. The user can control the amount of smoothing by setting the \code{bandwidth} parameter. The larger parameter, the smoother the series.
An infinitesimal bandwidth amounts to no smoothing.\cr

\code{donors} must be a character vector with one or more champions contained in \code{dta}. Alternatively, one of seven special strings can be used:

\itemize{
   \item{"all" }{This includes all champions.}
   \item{"non_lgb" }{This includes all champions expect Graves, Nami, Leona, Diana, and Neeko.}
   \item{"jungle" }{This includes all champions whose main role is Jungle.}
   \item{"middle" }{This includes all champions whose main role is Middle.}
   \item{"top" }{This includes all champions whose main role is Top.}
   \item{"support" }{This includes all champions expect whose main role is Support.}
   \item{"adc" }{This includes all champions expect whose main role is Adc.}
}

It is possible to include \code{"LGB"} in \code{champions}. If so, \code{\link{run_main_pooled}} constructs a new unit by averaging the outcomes of Nami, Leona, Diana, and Neeko and
runs the analysis detailed above on this new unit. This is compatible only with \code{donors} set to \code{"non_lgb"}.
}
\seealso{
\code{\link{run_main_regional}}
}
\author{
Riccardo Di Francesco
}
